+++ LevelSystem/Runtime/Nodes/LevelNode.cs
@@ -1,0 +1,11 @@
+using Core.Nodes;
+using StatSystem;
+
+namespace LevelSystem.Nodes
+{
+    public class LevelNode : CodeFunctionNode
+    {
+        public ILevelable levelable;
+        public override float value => levelable.level;
+    }
+}